+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2026-02-26
+
+### Added
+- Implemented a custom application icon for all platforms.
+- Added a native splash screen (Android, iOS) for a seamless startup experience.
+- Created a dynamic, fullscreen Flutter splash screen that shows initialization progress (loading preferences, starting backends, etc.).
+
 ## [1.1.0] - 2026-02-26
 
 ### Added